requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
playwright>=1.40.0
openai>=1.3.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.24.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.25.0
lxml>=4.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
